src/api-processing/pyproject.toml
[project]
name = "api-processing"
version = "0.1.0"
description = "FastAPI service for image processing: uploads to Blob and enqueues message to Service Bus"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [ { name = "Demo" } ]
dependencies = [
  "fastapi",
  "uvicorn",
  "python-dotenv",
  "azure-identity",
  "azure-storage-blob",
  "azure-servicebus",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "azure-monitor-opentelemetry",
  "azure-core-tracing-opentelemetry",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-requests",
  "python-multipart",
  "aiohttp",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-asyncio",
  "httpx",
  "asgi-lifespan",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "pytest-asyncio",
  "httpx",
  "asgi-lifespan",
]
